refactor(shipping): Phase 2a move integration event payloads to Contracts/
- Move 8 payloads (ShipmentCreated/Dispatched/Delivered/Cancelled/Failed/
  Returned/StatusChanged + consumed OrderConfirmed) to Contracts/Integration/
  under Shipping.Service.Contracts.Integration; handlers stay in
  IntegrationEvents/EventHandlers/ until Phase 7 dissolves them.
- Bump ECommerce.Shared 2.23.0 -> 2.25.0 to clear the pre-existing
  MessagingProviderBootTests RabbitMQ failure (latent eager-connection
  defect documented in db7977c); required to satisfy Phase 2a's
  "tests green" criterion.
