+++ shipping-microservice/Shipping.Service/Carriers/CarrierStatusApplier.cs
@@ -1,5 +1,5 @@
 using ECommerce.Shared.Infrastructure.EventBus;
-using Shipping.Service.IntegrationEvents;
+using Shipping.Service.Contracts.Integration;
 using Shipping.Service.Models;
 using Shipping.Service.Observability;
