+++ shipping-microservice/Shipping.Service/Contracts/Integration/OrderConfirmedEvent.cs
@@ -1,5 +1,5 @@
 using ECommerce.Shared.Infrastructure.EventBus;
 
-namespace Shipping.Service.IntegrationEvents;
+namespace Shipping.Service.Contracts.Integration;
 
 public record OrderConfirmedEvent(Guid OrderId, string CustomerId) : Event;
