+++ shipping-microservice/Shipping.Service/Contracts/Integration/ShipmentCancelledEvent.cs
@@ -1,6 +1,6 @@
 using ECommerce.Shared.Infrastructure.EventBus;
 
-namespace Shipping.Service.IntegrationEvents;
+namespace Shipping.Service.Contracts.Integration;
 
 public record ShipmentCancelledEvent(
     Guid ShipmentId,
