+++ shipping-microservice/Shipping.Service/Contracts/Integration/ShipmentCreatedEvent.cs
@@ -1,6 +1,6 @@
 using ECommerce.Shared.Infrastructure.EventBus;
 
-namespace Shipping.Service.IntegrationEvents;
+namespace Shipping.Service.Contracts.Integration;
 
 public record ShipmentLineItem(int ProductId, int Quantity);
