+++ shipping-microservice/Shipping.Service/Contracts/Integration/ShipmentDeliveredEvent.cs
@@ -1,12 +1,11 @@
 using ECommerce.Shared.Infrastructure.EventBus;
 
-namespace Shipping.Service.IntegrationEvents;
+namespace Shipping.Service.Contracts.Integration;
 
-public record ShipmentFailedEvent(
+public record ShipmentDeliveredEvent(
     Guid ShipmentId,
     Guid OrderId,
     string CustomerId,
     string? CarrierKey,
     string? TrackingNumber,
-    string Reason,
     DateTime OccurredAt) : Event;
