+++ shipping-microservice/Shipping.Service/Contracts/Integration/ShipmentDispatchedEvent.cs
@@ -1,6 +1,6 @@
 using ECommerce.Shared.Infrastructure.EventBus;
 
-namespace Shipping.Service.IntegrationEvents;
+namespace Shipping.Service.Contracts.Integration;
 
 public record ShipmentDispatchedEvent(
     Guid ShipmentId,
