+++ shipping-microservice/Shipping.Service/Shipping.Service.csproj
@@ -9,7 +9,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="ECommerce.Shared" Version="2.23.0" />
+    <PackageReference Include="ECommerce.Shared" Version="2.25.0" />
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="10.0.5" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="10.0.5">
       <PrivateAssets>all</PrivateAssets>
