+++ shipping-microservice/Shipping.Tests/Api/CancelShipmentCommandHandlerTests.cs
@@ -2,7 +2,7 @@ using System.Text.Json;
 using ECommerce.Shared.Infrastructure.Outbox;
 using ECommerce.Shared.IntegrationEvents.Commands;
 using Microsoft.Extensions.DependencyInjection;
-using Shipping.Service.IntegrationEvents;
+using Shipping.Service.Contracts.Integration;
 using Shipping.Service.IntegrationEvents.EventHandlers;
 using Shipping.Service.Models;
