+++ shipping-microservice/Shipping.Tests/Api/ShipmentDispatchEndpointsTests.cs
@@ -4,7 +4,7 @@ using ECommerce.Shared.Infrastructure.Outbox;
 using Microsoft.Extensions.DependencyInjection;
 using Shipping.Service.ApiModels;
 using Shipping.Service.Carriers;
-using Shipping.Service.IntegrationEvents;
+using Shipping.Service.Contracts.Integration;
 using Shipping.Service.Models;
 using Shipping.Tests.Authentication;
