+++ shipping-microservice/Shipping.Tests/Api/ShipmentTerminalTransitionTests.cs
@@ -3,7 +3,7 @@ using System.Net.Http.Json;
 using ECommerce.Shared.Infrastructure.Outbox;
 using Microsoft.Extensions.DependencyInjection;
 using Shipping.Service.ApiModels;
-using Shipping.Service.IntegrationEvents;
+using Shipping.Service.Contracts.Integration;
 using Shipping.Service.Models;
 using Shipping.Tests.Authentication;
