+++ shipping-microservice/Shipping.Tests/Api/ShipmentWebhookTests.cs
@@ -4,7 +4,7 @@ using ECommerce.Shared.Infrastructure.Outbox;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Shipping.Service.Carriers;
-using Shipping.Service.IntegrationEvents;
+using Shipping.Service.Contracts.Integration;
 using Shipping.Service.Models;
 
 namespace Shipping.Tests.Api;
