+++ shipping-microservice/Shipping.Tests/Carriers/CarrierPollingServiceTests.cs
@@ -2,7 +2,7 @@ using ECommerce.Shared.Infrastructure.Outbox;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Shipping.Service.Carriers;
-using Shipping.Service.IntegrationEvents;
+using Shipping.Service.Contracts.Integration;
 using Shipping.Service.Models;
 
 namespace Shipping.Tests.Carriers;
