+++ shipping-microservice/Shipping.Tests/IntegrationEvents/EventStreamConsolidationTests.cs
@@ -8,8 +8,8 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Shipping.Service.ApiModels;
 using Shipping.Service.Carriers;
+using Shipping.Service.Contracts.Integration;
 using Shipping.Service.Infrastructure.Data.EntityFramework;
-using Shipping.Service.IntegrationEvents;
 using Shipping.Service.Models;
 
 namespace Shipping.Tests.IntegrationEvents;
